Allow up to 65535 shards (#469)
Fixes #462

+++ src/shard.c
@@ -24,7 +24,7 @@ static uint32_t shard_roll_to_valid(shard_t *s)
 	return shard_get_next_ip(s);
 }
 
-void shard_init(shard_t *shard, uint8_t shard_idx, uint8_t num_shards,
+void shard_init(shard_t *shard, uint16_t shard_idx, uint16_t num_shards,
 		uint8_t thread_idx, uint8_t num_threads,
 		uint32_t max_total_targets, const cycle_t *cycle,
 		shard_complete_cb cb, void *arg)
