+++ src/shard.h
@@ -39,7 +39,7 @@ typedef struct shard {
 	void *arg;
 } shard_t;
 
-void shard_init(shard_t *shard, uint8_t shard_idx, uint8_t num_shards,
+void shard_init(shard_t *shard, uint16_t shard_idx, uint16_t num_shards,
 		uint8_t thread_idx, uint8_t num_threads,
 		uint32_t max_total_targets, const cycle_t *cycle,
 		shard_complete_cb cb, void *arg);
